It was a bad idea to model a trunk as a state machine.
Rewrite to make it command-driven instead of state driven.

This simplifies & clarifies the code, but eventually, we need to treat
calls (or rather call legs) individually.

One remaining problem is that a disconnectRequest and a disconnect may cross,
and if that happens, a disconnect will not be acknowledged (with an error).


git-svn-id: file:///Library/svn/voice3@96 da93438f-0fd1-0310-b142-832b9e69d951