The Villa was having rare sequencer crashes, trying to log a molecule that looked like it had been deleted already.
The only smoking gun was that extremely short files were played.

I couldn't reproduce the problem, but at least in theory, Sequencer::completed could be called recursively and this 
could delete a molecule while the lower function on the stack thought it was just started.

As a workaround, avoid recursive calls to ::completed. It's a bit hackish, but not too much.


git-svn-id: file:///Library/svn/voice3@253 da93438f-0fd1-0310-b142-832b9e69d951